Second version, added react frontend

+++ src/main/java/com/emilianokloster/kry/domain/EndpointService.java
@@ -1,5 +1,6 @@
 package com.emilianokloster.kry.domain;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -8,7 +9,9 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.emilianokloster.kry.entities.Endpoint;
+import com.emilianokloster.kry.model.BasicEndpoint;
+import com.emilianokloster.kry.model.Endpoint;
+import com.emilianokloster.kry.model.EndpointRepository;
 
 @Service
 public class EndpointService {
@@ -16,44 +19,66 @@ public class EndpointService {
 	@Autowired
 	private EndpointRepository repository;
 	
-	private List<Endpoint> endpoints = new ArrayList<>();
+	private List<Endpoint> endpoints;
 	
+	/*
+	 * This method is called repeatedly by the polling service.
+	 * It's configured to fetch data from the database just once, otherwise, 
+	 * it would be too expensive. It fills a list with all the values and that 
+	 * collection is constantly updated to provide the url's to validate. 
+	 */
 	public List<Endpoint> getAll() {
-		endpoints = new ArrayList<>(); 
-		repository.findAll().forEach(endpoints::add);
+		if (endpoints == null) {
+			endpoints = new ArrayList<>(); 
+			repository.findAll().forEach(endpoints::add);
+		}
 		return endpoints;
 	}
 	
-	public Optional<Endpoint> get(String id) {
-		return repository.findById(id);
-		
-//		return endpoints.stream()
-//				.filter(ep -> ep.getId().equals(id))
-//				.findFirst().get();
+	public List<BasicEndpoint> getAllBasic() {
+		// TODO Implement to be used by polling service
+		return null;
+	}
+	
+	public Endpoint get(Long id) {
+		if (id == null) return null;
+		return endpoints.stream()
+				.filter(ep -> ep.getId().equals(id))
+				.findFirst().get();
 	}
 
 	public void add(Endpoint endpoint) {
-		repository.save(endpoint);
-		
-//		endpoints.add(endpoint);
+		if (endpoint != null) {
+			endpoint.setDateInsert(LocalDate.now());
+			endpoint.setDateLastUpdate(LocalDate.now());
+			
+			// TODO endpoint.setUserInsert()
+			
+			final Endpoint newInstance = repository.save(endpoint);
+			endpoints.add(newInstance);
+		}
 	}
 
-	public void update(Endpoint endpoint, String id) {
-		repository.save(endpoint);
-		
-//		endpoints.stream()
-//			.filter(ep -> ep.getId().equals(id))
-//			.findFirst()
-//			.ifPresent(ep -> {
-//				ep.setName(endpoint.getName());
-//				ep.setUrl(endpoint.getUrl());
-//			});
+	public void update(Endpoint endpoint, Long id) {
+		if (endpoint != null && id != null) {
+			final Endpoint updated = repository.save(endpoint);
+			endpoints.stream()
+				.filter(ep -> ep.getId().equals(id))
+				.findFirst()
+				.ifPresent(ep -> {
+					ep.setName(updated.getName());
+					ep.setUrl(updated.getUrl());
+					ep.setComment(updated.getComment());
+					ep.setDateLastUpdate(LocalDate.now());
+				});
+		}
 	}
 
-	public void delete(String id) {
-		repository.deleteById(id);
-		
-//		endpoints.removeIf(t -> t.getId().equals(id));
+	public void delete(Long id) {
+		if (id != null && !id.equals("")) {
+			repository.deleteById(id);
+			endpoints.removeIf(ep -> ep.getId().equals(id));
+		}
 	}
 
 }
